[amdgcn] Move sched attr impls to AMDGCNScheduler to fix layering
SchedAttrs.cpp inside AMDGCNDialect included WaitAnalysis.h from
AMDGCNAnalysis, which depends on AMDGCNDialect — a cycle. Remove
DeclareAttrInterfaceMethods from the five scheduling attributes and
implement them as FallbackModel external models in a new AMDGCNScheduler
library (Analysis depends on Dialect; Scheduler depends on Analysis).
Register via registerAMDGCNSchedulerExternalModels in Init.cpp.
