+++ lib/Dialect/AMDGCN/Scheduler/CMakeLists.txt
@@ -1,0 +1,16 @@
+add_mlir_library(AMDGCNScheduler
+  GraphBuilderAttrs.cpp
+  SchedulerAttrs.cpp
+
+  DEPENDS
+  SchedInterfacesIncGen
+
+  LINK_LIBS PUBLIC
+  AMDGCNDialect
+  AMDGCNAnalysis
+  AsterInterfaces
+  LLVMSupport
+  MLIRAnalysis
+  MLIRIR
+  MLIRSupport
+)
